replace vec3 with crate, structure camera

+++ src/utility/hitlist.rs
@@ -15,7 +15,7 @@ impl<T: Hittable> HitList<T> {
 }
 
 impl<T: Hittable> Hittable for HitList<T> {
-    fn hit(&self, ray: &Ray, t_min: f64, t_max: f64, hit_record: &mut HitRecord) -> bool {
+    fn hit(&self, ray: &Ray, t_min: f32, t_max: f32, hit_record: &mut HitRecord) -> bool {
         let mut temp_rec = HitRecord::default();
         let mut hit_anything = false;
         let mut closest_so_far = t_max;
