+++ src/utility/hittable.rs
@@ -1,5 +1,5 @@
 use super::{ray::Ray, hitrecord::HitRecord};
 
 pub trait Hittable {
-    fn hit(&self, ray: &Ray, t_min: f64, t_max: f64, hit_record: &mut HitRecord) -> bool;
+    fn hit(&self, ray: &Ray, t_min: f32, t_max: f32, hit_record: &mut HitRecord) -> bool;
 }
