+++ src/utility/ray.rs
@@ -1,4 +1,4 @@
-use super::vec3::Vec3;
+use glam::Vec3;
 
 pub struct Ray {
     pub origin: Vec3,
@@ -6,7 +6,7 @@ pub struct Ray {
 }
 
 impl Ray {
-    pub fn at(&self, t: f64) -> Vec3 {
+    pub fn at(&self, t: f32) -> Vec3 {
         self.origin + self.direction * t
     }
 }
